+++ src/job-handlers/worker.ts
@@ -1,5 +1,9 @@
+import express from 'express'
+import { createBullBoard } from '@bull-board/api'
+import { BullAdapter } from '@bull-board/api/bullAdapter'
+import { ExpressAdapter } from '@bull-board/express'
 import connection from '../connection'
-import { collectionStats } from '../jobs/collection-stats'
+import { collectionStatsQueue } from '../jobs/collection-stats'
 import { metadataQueue } from '../jobs/process-metadata'
 import metadataHandler from './process-metadata'
 import collectionStatsHandler from './collection-stats'
@@ -15,7 +19,36 @@ async function main() {
     console.info('handling jobs...')
 
     metadataQueue.process(50, metadataHandler)
-    collectionStats.process(20, collectionStatsHandler)
+    collectionStatsQueue.process(20, collectionStatsHandler)
+
+    const serverAdapter = new ExpressAdapter()
+    serverAdapter.setBasePath('/')
+
+    createBullBoard({
+        queues: [new BullAdapter(collectionStatsQueue), new BullAdapter(metadataQueue)],
+        serverAdapter,
+        options: {
+            uiConfig: {
+                boardTitle: 'Indexer Queue',
+                boardLogo: {
+                    path: 'https://cdn.nft.io/branding/enjin.svg',
+                    width: '50px',
+                    height: 80,
+                },
+            },
+        },
+    })
+
+    const app = express()
+
+    app.use('/', serverAdapter.getRouter())
+
+    // other configurations of your server
+
+    app.listen(6000, () => {
+        // eslint-disable-next-line no-console
+        console.log('Running on 6000...')
+    })
 }
 
 main()
